Keep the RFCOMM tty fd open, so that MM won't keep reconnecting
See: https://bugzilla.gnome.org/show_bug.cgi?id=736984

For Bluez4 the bluetoothd does this as we call Connect via DBus.

NB: Despite it's not wise anyway, there's a bug in the kernel that deferred
rfcomm tty teardown races with the new open. The newly open tty gets a HUP
which confuses NM and the deallocation fails on a refcount check, making the
old rfcomm stuck in the kernel, needing a reboot for another reopen to succeed.