libnm: drop unused NMRemoteConnectionError
NMRemoteConnection used to return
NM_REMOTE_CONNECTION_ERROR_DISCONNECTED if you tried to operate on a
connection that had been disconnected from its D-Bus proxy. But this
disappeared in the gdbus port (since gdbus doesn't emit a signal when
it happens, so it's harder to notice. And it's not clear why
NMRemoteConnection did this when no other class did anyway...).